-- V1.16 -- - Base yapıya updated eklendi. - Bazı modellere eksik özellikler eklendi. - Ekleme işlemlerinde oluşturulma ve güncelleme Interceptor haline getirildi.

+++ E-ticaret/Server/Core/ETicaretAPI.Domain/Entities/Common/Product.cs
@@ -9,7 +9,7 @@ namespace ETicaretAPI.Domain.Entities.Common
     {
         public string Name { get; init; } = null!;
         public int Stock { get; init; }
-        public long Price { get; init; }
+        public float Price { get; init; }
         public ICollection<Order>? Orders { get; init; }
     }
 }
